app/requirements.txt
Flask==0.12.3
mysql-connector
